reorganize modules to avoid the ambiguous unchecked methods
