+++ src/sparse/linalg/trisolve.rs
@@ -4,6 +4,7 @@ use std::ops::IndexMut;
 use num_traits::Num;
 use sparse::CsMatView;
 use sparse::vec;
+use sparse::CsVecView;
 use errors::SprsError;
 use stack::{self, StackVal, DStack};
 
@@ -103,7 +104,7 @@ where N: Copy + Num,
 }
 
 fn lspsolve_csc_process_col<N: Copy + Num, V: ?Sized>
-                                                      (col: vec::CsVecView<N>,
+                                                      (col: CsVecView<N>,
                                                        col_ind: usize,
                                                        rhs: &mut V)
                                                        -> Result<(), SprsError>
@@ -249,7 +250,7 @@ where N: Copy + Num,
 /// * if w_workspace is not of length n
 ///
 pub fn lsolve_csc_sparse_rhs<N>(lower_tri_mat: CsMatView<N>,
-                                rhs: vec::CsVecView<N>,
+                                rhs: CsVecView<N>,
                                 dstack: &mut DStack<StackVal<usize>>,
                                 x_workspace: &mut [N],
                                 visited: &mut [bool]
@@ -318,7 +319,7 @@ where N: Copy + Num
 #[cfg(test)]
 mod test {
 
-    use sparse::{CsMatOwned, vec};
+    use sparse::{CsMatOwned, CsVecOwned};
     use stack::{self, DStack};
     use std::collections::HashSet;
 
@@ -397,7 +398,7 @@ mod test {
                                     vec![0, 2, 5, 6, 8, 9],
                                     vec![0, 1, 1, 2, 4, 2, 3, 4, 4],
                                     vec![1, 1, 2, 3, 2, 3, 7, 3, 5]);
-        let b = vec::CsVecOwned::new(5, vec![1, 2, 4], vec![4, 9, 9]);
+        let b = CsVecOwned::new(5, vec![1, 2, 4], vec![4, 9, 9]);
         let mut xw = vec![1; 5]; // inital values should not matter
         let mut visited = vec![false; 5]; // inital values matter here
         let mut dstack = DStack::with_capacity(2 * 5);
@@ -413,9 +414,9 @@ mod test {
                                   .map(|&i| (i, xw[i]))
                                   .collect();
 
-        let expected_output = vec::CsVecOwned::new(5,
-                                                   vec![1, 2, 4],
-                                                   vec![2, 1, 1]);
+        let expected_output = CsVecOwned::new(5,
+                                              vec![1, 2, 4],
+                                              vec![2, 1, 1]);
         let expected_output = expected_output.to_set();
 
         assert_eq!(x, expected_output);
@@ -433,9 +434,9 @@ mod test {
                                          5, 6],
                                     vec![1, 1, 2, 3, 3, 1, 7, 5, 2,
                                          1, 2]);
-        let b = vec::CsVecOwned::new(7,
-                                     vec![0, 2, 3, 5],
-                                     vec![1, 7, 7, 3]);
+        let b = CsVecOwned::new(7,
+                                vec![0, 2, 3, 5],
+                                vec![1, 7, 7, 3]);
         let mut dstack = DStack::with_capacity(2 * 7);
         let mut xw = vec![1; 7]; // inital values should not matter
         let mut visited = vec![false; 7]; // inital values matter here
@@ -451,10 +452,10 @@ mod test {
                                   .map(|&i| (i, xw[i]))
                                   .collect();
 
-        let expected_output = vec::CsVecOwned::new(7,
-                                                   vec![0, 2, 3, 5],
-                                                   vec![1, 2, 1, 1]
-                                                  ).to_set();
+        let expected_output = CsVecOwned::new(7,
+                                              vec![0, 2, 3, 5],
+                                              vec![1, 2, 1, 1]
+                                             ).to_set();
 
         assert_eq!(x, expected_output);
     }
